# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Engineering Gift API - ELV Calculations platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Calculation {
  id: number;
  /** Type of calculation: voltage_drop, fiber_budget, inrush_current */
  type: string;
  projectName?: string;
  engineerName?: string;
  /** JSON string of input values */
  inputs: string;
  /** JSON string of result values */
  results: string;
  createdAt: string;
}

export interface CreateCalculationInput {
  type: string;
  projectName?: string;
  engineerName?: string;
  inputs: string;
  results: string;
}

export interface DeleteResult {
  success: boolean;
}