# lib/api-zod/src/generated/types/deleteResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Engineering Gift API - ELV Calculations platform
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteResult {
  success: boolean;
}